refactor: Cache 클래스 support/cache 이동
