+++ src/main/java/net/detalk/api/plan/service/PricingPlanService.java
@@ -3,7 +3,7 @@ package net.detalk.api.plan.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.detalk.api.plan.domain.PricingPlan;
-import net.detalk.api.service.DetalkCache;
+import net.detalk.api.support.cache.DetalkCache;
 import org.springframework.stereotype.Service;
 
 @Slf4j
