+++ src/test/java/net/detalk/api/mock/FakePricingPlanCache.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import net.detalk.api.plan.domain.PricingPlan;
 import net.detalk.api.plan.domain.exception.PricingPlanNotFoundException;
-import net.detalk.api.service.DetalkCache;
+import net.detalk.api.support.cache.DetalkCache;
 
 public class FakePricingPlanCache implements DetalkCache<String, PricingPlan> {
